Good

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 //Video.find({},  (error, videos) => {});
 
@@ -8,7 +9,8 @@ export const home = async (req, res) => {
 };
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
+    console.log(video);
     if(!video) {
         return res.render("404", {pageTitle: "Video not found"});
     } 
@@ -16,18 +18,30 @@ export const watch = async (req, res) => {
 };
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const {user:{
+        _id,
+    }} = req.session;
     const video = await Video.findById(id);
     if(!video){
         return res.render("404", {pageTitle: "Video not found"});
+    }
+    if(String(video.owner) !== _id){
+        return res.status(403).redirect("/");
     }
     return res.render("edit", {pageTitle : `Edit ${video.title}`, video});
 };
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const {user:{
+        _id,
+    }} = req.session;
     const {title, description, hashtags} = req.body;
     const video = await Video.exists({_id: id});
     if(!video){
         return res.render("404", {pageTitle: "Video not found"});
+    }
+    if(String(video.owner) !== _id){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {title, description, hashtags: Video.formatHashtags(hashtags)});
     return res.redirect(`/videos/${id}`);
@@ -36,13 +50,24 @@ export const getUpload = (req, res) => {
     return res.render("upload", {pageTitle: "Upload Video"});
 };
 export const postUpload = async (req, res) => {
+    const {
+        user: {
+            _id
+        }
+    } = req.session;
+    const { path : fileUrl } = req.file;
     const {title, description, hashtags} = req.body
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags)
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     }catch(error) {
         return res.status(400).render("upload", {pageTitle: "Upload Video", errorMessage: error._message});
@@ -52,6 +77,13 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const {id} = req.params;
+    const {user:{
+        _id,
+    }} = req.session;
+    const video = await Video.findById(id);
+    if(String(video.owner) !== _id){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
